# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXUS COMM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  online: 'online',
  offline: 'offline',
  away: 'away',
} as const;

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  status: UserStatus;
  role: UserRole;
  createdAt: string;
}

export type UserStatusUpdateStatus = typeof UserStatusUpdateStatus[keyof typeof UserStatusUpdateStatus];


export const UserStatusUpdateStatus = {
  online: 'online',
  offline: 'offline',
  away: 'away',
} as const;

export interface UserStatusUpdate {
  status: UserStatusUpdateStatus;
}

export interface Room {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  isPrivate: boolean;
  createdBy: number;
  memberCount: number;
  /** @nullable */
  lastMessageAt?: string | null;
  createdAt: string;
}

export interface RoomInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  isPrivate?: boolean;
}

export interface RoomMember {
  id: number;
  roomId: number;
  userId: number;
  user?: User;
  joinedAt: string;
}

export type MessageType = typeof MessageType[keyof typeof MessageType];


export const MessageType = {
  text: 'text',
  image: 'image',
  file: 'file',
  system: 'system',
} as const;

export interface Message {
  id: number;
  roomId: number;
  userId: number;
  content: string;
  type: MessageType;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  mimeType?: string | null;
  user?: User;
  createdAt: string;
}

export type MessageInputType = typeof MessageInputType[keyof typeof MessageInputType];


export const MessageInputType = {
  text: 'text',
  image: 'image',
  file: 'file',
  system: 'system',
} as const;

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  type?: MessageInputType;
  fileUrl?: string;
  fileName?: string;
  mimeType?: string;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  message: 'message',
  call: 'call',
  system: 'system',
} as const;

export interface Notification {
  id: number;
  userId: number;
  type: NotificationType;
  title: string;
  /** @nullable */
  content?: string | null;
  isRead: boolean;
  /** @nullable */
  relatedUserId?: number | null;
  /** @nullable */
  relatedRoomId?: number | null;
  createdAt: string;
}

export type CallCallType = typeof CallCallType[keyof typeof CallCallType];


export const CallCallType = {
  audio: 'audio',
  video: 'video',
} as const;

export type CallStatus = typeof CallStatus[keyof typeof CallStatus];


export const CallStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  missed: 'missed',
  completed: 'completed',
} as const;

export interface Call {
  id: number;
  initiatorId: number;
  recipientId: number;
  callType: CallCallType;
  status: CallStatus;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  roomId?: number | null;
  startedAt?: string;
  /** @nullable */
  endedAt?: string | null;
  createdAt: string;
}

export type CallInputCallType = typeof CallInputCallType[keyof typeof CallInputCallType];


export const CallInputCallType = {
  audio: 'audio',
  video: 'video',
} as const;

export interface CallInput {
  recipientId: number;
  callType: CallInputCallType;
  roomId?: number;
}

export type CallStatusUpdateStatus = typeof CallStatusUpdateStatus[keyof typeof CallStatusUpdateStatus];


export const CallStatusUpdateStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  missed: 'missed',
  completed: 'completed',
} as const;

export interface CallStatusUpdate {
  status: CallStatusUpdateStatus;
  duration?: number;
}

export interface ActivityItem {
  type: string;
  description: string;
  createdAt: string;
}

export interface DashboardStats {
  onlineUsers: number;
  totalUsers: number;
  totalRooms: number;
  totalMessages: number;
  recentActivity: ActivityItem[];
}