# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NEXUS COMM API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: rooms
    description: Chat rooms
  - name: messages
    description: Chat messages
  - name: notifications
    description: Notifications
  - name: calls
    description: Call history
  - name: stats
    description: Dashboard stats

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

  /users/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      summary: Update current user status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all chat rooms
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a chat room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

  /rooms/{id}/join:
    post:
      operationId: joinRoom
      tags: [rooms]
      summary: Join a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /rooms/{id}/members:
    get:
      operationId: getRoomMembers
      tags: [rooms]
      summary: Get room members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomMember"

  /rooms/{id}/messages:
    get:
      operationId: getRoomMessages
      tags: [messages]
      summary: Get messages in a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get my notifications
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /calls:
    post:
      operationId: initiateCall
      tags: [calls]
      summary: Initiate a call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallInput"
      responses:
        "201":
          description: Call created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"

  /calls/{id}/status:
    patch:
      operationId: updateCallStatus
      tags: [calls]
      summary: Update call status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /stats/dashboard:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, offline, away]
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
      required: [id, username, status, role, createdAt]

    UserStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [online, offline, away]
      required: [status]

    Room:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        isPrivate:
          type: boolean
        createdBy:
          type: integer
        memberCount:
          type: integer
        lastMessageAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, name, isPrivate, createdBy, memberCount, createdAt]

    RoomInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        isPrivate:
          type: boolean
          default: false
      required: [name]

    RoomMember:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        joinedAt:
          type: string
          format: date-time
      required: [id, roomId, userId, joinedAt]

    Message:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        userId:
          type: integer
        content:
          type: string
        type:
          type: string
          enum: [text, image, file, system]
        fileUrl:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required: [id, roomId, userId, content, type, createdAt]

    MessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        type:
          type: string
          enum: [text, image, file, system]
          default: text
        fileUrl:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
      required: [content]

    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [message, call, system]
        title:
          type: string
        content:
          type: ["string", "null"]
        isRead:
          type: boolean
        relatedUserId:
          type: ["integer", "null"]
        relatedRoomId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, title, isRead, createdAt]

    Call:
      type: object
      properties:
        id:
          type: integer
        initiatorId:
          type: integer
        recipientId:
          type: integer
        callType:
          type: string
          enum: [audio, video]
        status:
          type: string
          enum: [pending, accepted, rejected, missed, completed]
        duration:
          type: ["integer", "null"]
        roomId:
          type: ["integer", "null"]
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, initiatorId, recipientId, callType, status, createdAt]

    CallInput:
      type: object
      properties:
        recipientId:
          type: integer
        callType:
          type: string
          enum: [audio, video]
        roomId:
          type: integer
      required: [recipientId, callType]

    CallStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, accepted, rejected, missed, completed]
        duration:
          type: integer
      required: [status]

    DashboardStats:
      type: object
      properties:
        onlineUsers:
          type: integer
        totalUsers:
          type: integer
        totalRooms:
          type: integer
        totalMessages:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required: [onlineUsers, totalUsers, totalRooms, totalMessages, recentActivity]

    ActivityItem:
      type: object
      properties:
        type:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [type, description, createdAt]