# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXUS COMM API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  type: string;
  description: string;
  createdAt: Date;
}