# lib/api-zod/src/generated/types/callStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXUS COMM API
 * OpenAPI spec version: 0.1.0
 */
import type { CallStatusUpdateStatus } from './callStatusUpdateStatus';

export interface CallStatusUpdate {
  status: CallStatusUpdateStatus;
  duration?: number;
}