# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXUS COMM API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageInputType } from './messageInputType';

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  type?: MessageInputType;
  fileUrl?: string;
  fileName?: string;
  mimeType?: string;
}