# lib/api-zod/src/generated/types/roomMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXUS COMM API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface RoomMember {
  id: number;
  roomId: number;
  userId: number;
  user?: User;
  joinedAt: Date;
}